Arreglo Reporte

+++ JuncalApi/Modelos/JuncalProveedorCuentaCorriente.cs
@@ -19,7 +19,7 @@ public partial class JuncalProveedorCuentaCorriente
 
     public decimal? Importe { get; set; }
 
-    public decimal? Peso { get; set; }
+    public double? Peso { get; set; }
 
     public decimal? Total { get; set; }
